test for update offer helper
