+++ components/updater/card/src/test/kotlin/com/flipperdevices/updater/card/helpers/FileExistHelperTest.kt
@@ -1,9 +1,8 @@
-package com.flipperdevices.updater.card.utils
+package com.flipperdevices.updater.card.helpers
 
 import com.flipperdevices.bridge.api.manager.FlipperRequestApi
 import com.flipperdevices.protobuf.main
 import com.flipperdevices.protobuf.storage.md5sumResponse
-import com.flipperdevices.updater.card.helpers.FileExistHelper
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.flow.collectLatest
